aml-backend/requirements.prod.txt
Django==1.11.15
djangorestframework==3.6.4
drf-nested-routers==0.90.0
django-cors-headers==2.1.0
django-enumfield==1.2.1
django-extensions==1.9.1
django-extras==0.3
django-filter==1.0.4
django-redis==4.8.0
redis==2.10.6
mockredispy==2.9.3
django-rest-swagger==2.1.2
django-storages==1.6.5
boto3==1.4.7

Pillow==4.2.1
psycopg2==2.7.3.1

elasticsearch==6.3.0
gunicorn==19.7.1

PyYAML==3.12
msgpack-python==0.4.8

# celery
celery==4.1.0
amqp==2.2.2

# JWT
djangorestframework-jwt==1.11.0
